refactor: changed id from int to long

+++ src/main/java/com/noroff/noroffassignment_7/model/Franchise.java
@@ -8,7 +8,9 @@ import lombok.Setter;
 import javax.persistence.*;
 import javax.validation.constraints.Size;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Entity
@@ -18,7 +20,7 @@ public class Franchise {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Getter @Setter
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     @Getter @Setter
     @Size(min = 1, max = 45)
